refactor comment posting interface to separate component
