+++ components/UpdateComments.tsx
@@ -5,66 +5,23 @@ import {fetcher} from "../utils/utils";
 import axios from "axios";
 import {format} from "date-fns";
 import Link from "next/link";
+import {FiCornerUpRight} from "react-icons/fi";
+import UpdateCommentForm from "./UpdateCommentForm";
 
 export default function UpdateComments({updateId, userData}: { updateId: string, userData: User }) {
     const [refreshIteration, setRefreshIteration] = useState<number>(0);
     const {data, error} = useSWR(`/api/get-comments?updateId=${updateId}&?iter=${refreshIteration}`, fetcher);
-    const [commentText, setCommentText] = useState<string>("");
-    const [isLoading, setIsLoading] = useState<boolean>(false);
-
-    function postComment() {
-        setIsLoading(true);
-
-        axios.post("/api/new-comment", {
-            commentText: commentText,
-            authorId: userData._id,
-            updateId: updateId,
-        }).then(res => {
-            console.log(res);
-            setIsLoading(false);
-            clearComment();
-            setRefreshIteration(refreshIteration + 1)
-        }).catch(e => {
-            console.log(e);
-            setIsLoading(false);
-        });
-    }
-
-    function clearComment() {
-        setCommentText("");
-    }
 
     return (
         <>
             <div className="up-ui-title mb-4"><span>Comments {data ? `(${data.comments.length})` : ""}</span></div>
             <div className="my-4">
-                <textarea
-                    value={commentText}
-                    onChange={e => setCommentText(e.target.value)}
-                    className="w-full border p-4 rounded-md text-xl"
-                    placeholder="Write a comment..."
+                <UpdateCommentForm
+                    updateId={updateId}
+                    userData={userData}
+                    callback={() => setRefreshIteration(refreshIteration + 1)}
+                    onCancel={() => null}
                 />
-                <div className="flex mt-2">
-                    <div className="ml-auto relative">
-                        <button
-                            className="up-button text ml-auto mr-4"
-                            disabled={commentText.length === 0}
-                            onClick={clearComment}
-                        >
-                            Cancel
-                        </button>
-                        <button
-                            className="up-button primary small"
-                            disabled={commentText.length === 0 || isLoading}
-                            onClick={postComment}
-                        >
-                            Post
-                        </button>
-                        {isLoading && (
-                            <div className="up-spinner"/>
-                        )}
-                    </div>
-                </div>
             </div>
             <div className="my-4">
                 {data && data.comments.map((comment: CommentObj) => (
@@ -78,7 +35,7 @@ export default function UpdateComments({updateId, userData}: { updateId: string,
                                 />
                             </a>
                         </Link>
-                        <div className="ml-2">
+                        <div className="ml-2 w-full">
                             <p>
                                 <Link href={`/@${data.users.find(d => d._id === comment.authorId).urlName}`}>
                                     <a>
@@ -90,6 +47,10 @@ export default function UpdateComments({updateId, userData}: { updateId: string,
                                 </span>
                             </p>
                             <p className="text-xl">{comment.body}</p>
+                            <button className="opacity-25 hover:opacity-75 mt-2 inline-flex items-center">
+                                <FiCornerUpRight/>
+                                <span className="ml-2">Reply</span>
+                            </button>
                         </div>
                     </div>
                 ))}
